target - SortMergeJoinPerformer detail - Apply strategy pattern to SortMergeJoin perform reason - To change the behavior flexibly by separating the actual internal behavior into external algorithm objects
